+++ src/main/java/com/besson/endfield/screen/ModScreens.java
@@ -76,6 +76,10 @@ public class ModScreens {
     public static final ScreenHandlerType<FluidPumpScreenHandler> FLUID_PUMP_SCREEN =
             Registry.register(Registries.SCREEN_HANDLER, new Identifier(ArknightsEndfield.MOD_ID, "fluid_pump_screen"),
                     new ExtendedScreenHandlerType<>(FluidPumpScreenHandler::new));
+
+    public static final ScreenHandlerType<BigStorageScreenHandler> BIG_STORAGE_SCREEN =
+            Registry.register(Registries.SCREEN_HANDLER, new Identifier(ArknightsEndfield.MOD_ID, "big_storage_screen"),
+                    new ExtendedScreenHandlerType<>(BigStorageScreenHandler::new));
     public static void register() {
 
     }